web/django/requirements.txt
# requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
Django==3.1
pytz==2019.3
sqlparse==0.3.1
numpy==1.19.0
redis==4.3.1
redis-server==5.0.7
uwsgi